requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.11.8
distlib==0.3.1
Django==3.0.5
django-colorfield==0.3.2
django-cors-headers==3.5.0
djangorestframework==3.12.2
djongo==1.3.3
filelock==3.0.12
gunicorn==20.0.4
install==1.3.4
kafka-python==2.0.2
pip-chill==1.0.0
pipenv==2020.11.4
pippip==0.0.10
pymongo==3.11.0
pytz==2020.4
six==1.15.0
sqlparse==0.2.4
virtualenv==20.1.0
virtualenv-clone==0.5.4
